Fix: typoやデータ型の不整合を修正

+++ Database/Migrations/2024-03-31_1711875774_CreateCategoryTable.php
@@ -11,8 +11,8 @@ class CreateCategoryTable implements SchemaMigration
         // マイグレーション処理を書く
         return [
             "CREATE TABLE categories (
-                id INT PRIMARY KEY AUTO_INCREMENT,
-                category_name VARCHAR(255) NOT NULL,
+                id BIGINT PRIMARY KEY AUTO_INCREMENT,
+                category_name VARCHAR(255) NOT NULL
             )"
         ];
     }
